feat(plugins): extract irc tool to plugin
