+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## next
 
+### Improvements
+- **Extract IRC tool to plugin (`src/plugins/irc/`)** — moves IRC management and inspection tool (`irc`) out of core runtime tools into a modular plugin, keeping core tool definitions, scopes, and context assembly clean.
+
 ### Features
-- **Slack inspection and interaction tool (`slack`)** — built-in tool that allows the agent to inspect Slack workspaces on-demand without sending unsolicited messages: read channel history (`history`), read thread replies (`thread`), list public and private channels (`channels`), look up user profiles (`user`), list pinned messages (`pins`), view channel bookmarks (`bookmarks`), and add emoji reactions (`react`).
+- **Slack inspection and interaction tool (`slack`)** — built-in plugin that allows the agent to inspect Slack workspaces on-demand without sending unsolicited messages: read channel history (`history`), read thread replies (`thread`), list public and private channels (`channels`), look up user profiles (`user`), list pinned messages (`pins`), view channel bookmarks (`bookmarks`), and add emoji reactions (`react`).
 
 ## 0.35.0 (2026-09-10)
