+++ src/app.ts
@@ -16,7 +16,7 @@ import { registerAgent, writePidFile, removePidFile, assignPort } from "./regist
 import { AgentServer } from "./server.js";
 import { PairingManager } from "./pairing.js";
 import { setMessageSender } from "./tools/message.js";
-import { setIrcInterface, initIrcTool } from "./tools/irc.js";
+import { setIrcInterface } from "./plugins/irc/tools.js";
 import { SegmentIndex } from "./segments.js";
 import { MemoryDB } from "./memory.js";
 import { MessageQueue } from "./queue.js";
@@ -496,7 +496,6 @@ export async function startApp(agentDir: string, forceCli = false): Promise<void
   }
 
   // Start IRC if configured — IRC_URL overrides config.irc
-  initIrcTool(agentDir);
   const ircUrls = parseIrcUrls(process.env.IRC_URL || config.irc);
   let ircBot: IrcInterface | null = null;
   if (!forceCli && ircUrls.length) {
